# api/swagger.yaml
basePath: /
definitions:
...
  inbound.EmailVerifyRequest:
    properties:
      challenge_token:
        type: string
    type: object
...
  inbound.Login2FARequest:
    properties:
      challenge_token:
        type: string
      code:
        type: string
      method:
        type: string
    type: object
  inbound.Login2FAResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  inbound.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  inbound.LoginResponse:
    properties:
      access_token:
        type: string
      available_methods:
        items:
          type: string
        type: array
      challenge_token:
        type: string
      mfa_required:
        type: boolean
      refresh_token:
        type: string
    type: object
...
  inbound.PasswordForgotRequest:
    properties:
      email:
        type: string
    type: object
  inbound.PasswordResetRequest:
    properties:
      challenge_token:
        type: string
      new_password:
        type: string
    type: object
...
  inbound.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  inbound.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  inbound.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
  inbound.RegisterResendRequest:
    properties:
      email:
        type: string
    type: object
  inbound.RegisterResendResponse:
    type: object
...
  router.errorResponse:
    properties:
      error:
        additionalProperties:
          type: string
        type: object
      message:
        example: example string message
        type: string
    type: object
  router.successResponse:
    properties:
      data: {}
      message:
        example: example string message
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gobite.com
    name: Gobite API Support
    url: https://gobite.com/contact
  description: Gobite provides authentication and profile management APIs.
  license:
    name: MIT
    url: https://mit-license.org/
  termsOfService: https://gobite.com/terms
  title: Gobite API
  version: "1.0"
paths:
...
  /api/v1/identity/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns access/refresh tokens. If MFA
        is required, a challenge ID is returned.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication result
          schema:
            allOf:
            - $ref: '#/definitions/router.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/inbound.LoginResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Authenticate user
      tags:
      - Identity
  /api/v1/identity/login/2fa:
    post:
      consumes:
      - application/json
      description: Verifies the 2FA code for a login challenge and returns access/refresh
        tokens.
      parameters:
      - description: 2FA login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.Login2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication result
          schema:
            allOf:
            - $ref: '#/definitions/router.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/inbound.Login2FAResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Invalid MFA code
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Complete 2FA login
      tags:
      - Identity
...
  /api/v1/identity/password/forgot:
    post:
      consumes:
      - application/json
      description: Sends password reset instructions to the provided email address.
      parameters:
      - description: Forgot password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.PasswordForgotRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Request password reset
      tags:
      - Identity
  /api/v1/identity/password/reset:
    post:
      consumes:
      - application/json
      description: Sets a new password using the provided reset token.
      parameters:
      - description: Reset password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.PasswordResetRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "404":
          description: Reset token not found
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Reset password
      tags:
      - Identity
...
  /api/v1/identity/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new access/refresh token pair.
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refresh result
          schema:
            allOf:
            - $ref: '#/definitions/router.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/inbound.RefreshTokenResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Refresh access token
      tags:
      - Identity
  /api/v1/identity/register:
    post:
      consumes:
      - application/json
      description: Creates a new account and sends a verification email.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.RegisterRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Register user
      tags:
      - Identity
  /api/v1/identity/register/resend:
    post:
      consumes:
      - application/json
      description: Sends a new verification email when an account exists for the provided
        address.
      parameters:
      - description: Resend verification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.RegisterResendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Resend result
          schema:
            allOf:
            - $ref: '#/definitions/router.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/inbound.RegisterResendResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Resend verification email
      tags:
      - Identity
  /api/v1/identity/register/verify:
    post:
      consumes:
      - application/json
      description: Confirms the user's email address using the provided verification
        token.
      parameters:
      - description: Email verification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inbound.EmailVerifyRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/router.errorResponse'
        "404":
          description: Verification token not found
          schema:
            $ref: '#/definitions/router.errorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/router.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.errorResponse'
      summary: Verify email
      tags:
      - Identity
...
schemes:
- http
- https
swagger: "2.0"